Move BlockValueStore from sem_ir to base
