+++ toolchain/sem_ir/BUILD
@@ -7,20 +7,6 @@ load("//bazel/cc_rules:defs.bzl", "cc_library", "cc_test")
 package(default_visibility = ["//visibility:public"])
 
 cc_library(
-    name = "block_value_store",
-    hdrs = ["block_value_store.h"],
-    deps = [
-        "//common:check",
-        "//common:hashing",
-        "//common:set",
-        "//toolchain/base:mem_usage",
-        "//toolchain/base:value_store",
-        "//toolchain/base:yaml",
-        "@llvm-project//llvm:Support",
-    ],
-)
-
-cc_library(
     name = "typed_insts",
     srcs = [
         "ids.cpp",
@@ -124,7 +110,6 @@ cc_library(
         "builtin_function_kind.def",
     ],
     deps = [
-        ":block_value_store",
         ":clang_decl",
         ":typed_insts",
         "//common:check",
@@ -136,6 +121,7 @@ cc_library(
         "//common:raw_string_ostream",
         "//common:set",
         "//common:struct_reflection",
+        "//toolchain/base:block_value_store",
         "//toolchain/base:canonical_value_store",
         "//toolchain/base:index_base",
         "//toolchain/base:int",
