+++ toolchain/sem_ir/inst.h
@@ -13,10 +13,10 @@
 #include "common/ostream.h"
 #include "common/raw_string_ostream.h"
 #include "common/struct_reflection.h"
+#include "toolchain/base/block_value_store.h"
 #include "toolchain/base/index_base.h"
 #include "toolchain/base/int.h"
 #include "toolchain/base/value_store.h"
-#include "toolchain/sem_ir/block_value_store.h"
 #include "toolchain/sem_ir/id_kind.h"
 #include "toolchain/sem_ir/inst_kind.h"
 #include "toolchain/sem_ir/singleton_insts.h"
