+++ toolchain/sem_ir/struct_type_field.h
@@ -5,7 +5,7 @@
 #ifndef CARBON_TOOLCHAIN_SEM_IR_STRUCT_TYPE_FIELD_H_
 #define CARBON_TOOLCHAIN_SEM_IR_STRUCT_TYPE_FIELD_H_
 
-#include "toolchain/sem_ir/block_value_store.h"
+#include "toolchain/base/block_value_store.h"
 #include "toolchain/sem_ir/ids.h"
 
 namespace Carbon::SemIR {
